fix: updates per comments
