+++ api/test/unit/services/application-csv-export.service.spec.ts
@@ -2,7 +2,6 @@ import { randomUUID } from 'crypto';
 import { PassThrough } from 'stream';
 import { Test, TestingModule } from '@nestjs/testing';
 import {
-  ListingsStatusEnum,
   LotteryStatusEnum,
   MultiselectQuestionsApplicationSectionEnum,
 } from '@prisma/client';
@@ -29,9 +28,6 @@ import { Logger } from '@nestjs/common';
 import { SchedulerRegistry } from '@nestjs/schedule';
 import { GoogleTranslateService } from '../../../src/services/google-translate.service';
 import { CsvHeader } from '../../../src/types/CsvExportInterface';
-import { Application } from '../../../src/dtos/applications/application.dto';
-import MultiselectQuestion from '../../../src/dtos/multiselect-questions/multiselect-question.dto';
-import { OrderByEnum } from '../../../src/enums/shared/order-by-enum';
 
 describe('Testing application CSV export service', () => {
   let service: ApplicationCsvExporterService;
@@ -692,168 +688,6 @@ describe('Testing application CSV export service', () => {
     expect(readable).toContain('EST');
   });
 
-  describe('Testing lotteryRandomizerHelper()', () => {
-    it('should generate random ordinal arrays', () => {
-      for (let i = 0; i < 10; i++) {
-        // create and fill mock array
-        const filteredApplicationArray: Application[] = [];
-        for (let j = 0; j < i + 1; j++) {
-          filteredApplicationArray.push({
-            id: randomUUID(),
-          } as unknown as Application);
-        }
-
-        // run through randomizer
-        const ordinalArray = service.lotteryRandomizerHelper(
-          filteredApplicationArray,
-        );
-
-        // verify the results are random
-        expect(filteredApplicationArray.length).toEqual(ordinalArray.length);
-
-        for (let j = 1; j <= i + 1; j++) {
-          expect(ordinalArray.find((val) => val === j)).toEqual(j);
-        }
-      }
-    });
-
-    it('should store randomized ordinals when no preferences on listing', async () => {
-      const listingId = randomUUID();
-      const applications: Application[] = [];
-      for (let i = 0; i < 10; i++) {
-        applications.push({
-          id: randomUUID(),
-          markedAsDuplicate: false,
-        } as unknown as Application);
-      }
-
-      prisma.applicationLotteryPositions.createMany = jest
-        .fn()
-        .mockResolvedValue({ id: randomUUID() });
-
-      await service.lotteryRandomizer(listingId, applications, []);
-
-      expect(prisma.applicationLotteryPositions.createMany).toHaveBeenCalled();
-
-      const args = (prisma.applicationLotteryPositions.createMany as any).mock
-        .calls[0][0].data;
-
-      for (let i = 1; i < 11; i++) {
-        expect(args.find((val) => val.ordinal === i)).toEqual({
-          listingId,
-          applicationId: expect.anything(),
-          ordinal: i,
-          multiselectQuestionId: null,
-        });
-      }
-    });
-  });
-
-  describe('Testing lotteryRandomizer()', () => {
-    it('should store randomized ordinals when every application has a preference', async () => {
-      const listingId = randomUUID();
-      const applications: Application[] = [];
-      const preferences: MultiselectQuestion[] = [
-        {
-          id: randomUUID(),
-          text: 'example text',
-        } as unknown as MultiselectQuestion,
-      ];
-      for (let i = 0; i < 10; i++) {
-        applications.push({
-          id: randomUUID(),
-          markedAsDuplicate: false,
-          preferences: [
-            {
-              key: 'example text',
-              claimed: false,
-            },
-          ],
-        } as unknown as Application);
-      }
-
-      prisma.applicationLotteryPositions.createMany = jest
-        .fn()
-        .mockResolvedValue({ id: randomUUID() });
-
-      await service.lotteryRandomizer(listingId, applications, preferences);
-
-      const args = (prisma.applicationLotteryPositions.createMany as any).mock
-        .calls[0][0].data;
-
-      for (let i = 1; i < 11; i++) {
-        expect(args.find((val) => val.ordinal === i)).toEqual({
-          listingId,
-          applicationId: expect.anything(),
-          ordinal: i,
-          multiselectQuestionId: null,
-        });
-      }
-
-      expect(prisma.applicationLotteryPositions.createMany).toBeCalledTimes(1);
-    });
-
-    it('should store randomized ordinals and preference specific ordinals', async () => {
-      const listingId = randomUUID();
-      const applications: Application[] = [];
-      const preferences: MultiselectQuestion[] = [
-        {
-          id: randomUUID(),
-          text: 'example text',
-        } as unknown as MultiselectQuestion,
-      ];
-      for (let i = 0; i < 10; i++) {
-        applications.push({
-          id: randomUUID(),
-          markedAsDuplicate: false,
-          preferences: [
-            {
-              key: 'example text',
-              claimed: i % 2 === 0,
-            },
-          ],
-        } as unknown as Application);
-      }
-
-      prisma.applicationLotteryPositions.createMany = jest
-        .fn()
-        .mockResolvedValue({ id: randomUUID() });
-
-      await service.lotteryRandomizer(listingId, applications, preferences);
-
-      const args = (prisma.applicationLotteryPositions.createMany as any).mock
-        .calls[0][0].data;
-
-      for (let i = 1; i < 11; i++) {
-        expect(args.find((val) => val.ordinal === i)).toEqual({
-          listingId,
-          applicationId: expect.anything(),
-          ordinal: i,
-          multiselectQuestionId: null,
-        });
-      }
-
-      const argsWithPreference = (
-        prisma.applicationLotteryPositions.createMany as any
-      ).mock.calls[1][0].data;
-
-      for (let i = 1; i < 5; i++) {
-        expect(argsWithPreference.find((val) => val.ordinal === i)).toEqual({
-          listingId,
-          applicationId: expect.anything(),
-          ordinal: i,
-          multiselectQuestionId: expect.anything(),
-        });
-      }
-
-      expect(argsWithPreference.find((val) => val.ordinal === 6)).toEqual(
-        undefined,
-      );
-
-      expect(prisma.applicationLotteryPositions.createMany).toBeCalledTimes(2);
-    });
-  });
-
   describe('Testing lotteryExport()', () => {
     it('should generate lottery results ', async () => {
       process.env.TIME_ZONE = 'America/Los_Angeles';
@@ -921,199 +755,6 @@ describe('Testing application CSV export service', () => {
       });
 
       expect(readable).toContain('Lottery Position');
-    });
-  });
-
-  describe('Testing lotteryGenerate()', () => {
-    it('should build lottery when no prior lottery has been ran', async () => {
-      const listingId = randomUUID();
-      const requestingUser = {
-        firstName: 'requesting fName',
-        lastName: 'requesting lName',
-        email: 'requestingUser@email.com',
-        jurisdictions: [{ id: 'juris id' }],
-        userRoles: { isAdmin: true },
-      } as unknown as User;
-
-      permissionService.canOrThrow = jest.fn().mockResolvedValue(true);
-      prisma.listings.findUnique = jest.fn().mockResolvedValue({
-        id: listingId,
-        lotteryLastRunAt: null,
-        lotteryStatus: null,
-        status: ListingsStatusEnum.closed,
-      });
-      const applications = mockApplicationSet(5, new Date());
-      prisma.applications.findMany = jest.fn().mockReturnValue(applications);
-
-      prisma.multiselectQuestions.findMany = jest.fn().mockReturnValue([
-        {
-          ...mockMultiselectQuestion(
-            0,
-            new Date(),
-            MultiselectQuestionsApplicationSectionEnum.preferences,
-          ),
-          options: [
-            { id: 1, text: 'text' },
-            { id: 2, text: 'text', collectAddress: true },
-          ],
-        },
-        {
-          ...mockMultiselectQuestion(
-            1,
-            new Date(),
-            MultiselectQuestionsApplicationSectionEnum.programs,
-          ),
-          options: [{ id: 1, text: 'text' }],
-        },
-      ]);
-
-      prisma.applicationLotteryPositions.createMany = jest
-        .fn()
-        .mockResolvedValue({ id: randomUUID() });
-
-      prisma.listings.update = jest.fn().mockResolvedValue({
-        id: listingId,
-        lotteryLastRunAt: null,
-        lotteryStatus: null,
-      });
-
-      await service.lotteryGenerate(
-        { user: requestingUser } as unknown as ExpressRequest,
-        {} as unknown as Response,
-        { listingId },
-      );
-
-      expect(prisma.listings.findUnique).toHaveBeenCalledWith({
-        select: {
-          id: true,
-          lotteryLastRunAt: true,
-          lotteryStatus: true,
-        },
-        where: {
-          id: listingId,
-        },
-      });
-
-      expect(prisma.applications.findMany).toHaveBeenCalledWith({
-        select: {
-          id: true,
-          preferences: true,
-          householdMember: {
-            select: {
-              id: true,
-            },
-          },
-          applicationLotteryPositions: {
-            select: {
-              ordinal: true,
-              multiselectQuestionId: true,
-            },
-            where: {
-              multiselectQuestionId: null,
-            },
-            orderBy: {
-              ordinal: OrderByEnum.DESC,
-            },
-          },
-        },
-        where: {
-          listingId: listingId,
-          deletedAt: null,
-          markedAsDuplicate: false,
-        },
-      });
-
-      expect(prisma.applicationLotteryPositions.createMany).toHaveBeenCalled();
-      expect(prisma.listings.update).toHaveBeenCalledWith({
-        data: {
-          lotteryLastRunAt: expect.anything(),
-          lotteryStatus: LotteryStatusEnum.ran,
-        },
-        where: {
-          id: listingId,
-        },
-      });
-    });
-
-    it('should error when trying to generate lotterry results when previous results exist', async () => {
-      const listingId = randomUUID();
-      const requestingUser = {
-        firstName: 'requesting fName',
-        lastName: 'requesting lName',
-        email: 'requestingUser@email.com',
-        jurisdictions: [{ id: 'juris id' }],
-        userRoles: { isAdmin: true },
-      } as unknown as User;
-
-      permissionService.canOrThrow = jest.fn().mockResolvedValue(true);
-      prisma.listings.findUnique = jest.fn().mockResolvedValue({
-        id: listingId,
-        lotteryLastRunAt: new Date(),
-        lotteryStatus: LotteryStatusEnum.ran,
-      });
-      const applications = mockApplicationSet(5, new Date());
-      prisma.applications.findMany = jest.fn().mockReturnValue(applications);
-
-      prisma.multiselectQuestions.findMany = jest.fn().mockReturnValue([
-        {
-          ...mockMultiselectQuestion(
-            0,
-            new Date(),
-            MultiselectQuestionsApplicationSectionEnum.preferences,
-          ),
-          options: [
-            { id: 1, text: 'text' },
-            { id: 2, text: 'text', collectAddress: true },
-          ],
-        },
-        {
-          ...mockMultiselectQuestion(
-            1,
-            new Date(),
-            MultiselectQuestionsApplicationSectionEnum.programs,
-          ),
-          options: [{ id: 1, text: 'text' }],
-        },
-      ]);
-
-      prisma.applicationLotteryPositions.createMany = jest
-        .fn()
-        .mockResolvedValue({ id: randomUUID() });
-
-      prisma.listings.update = jest.fn().mockResolvedValue({
-        id: listingId,
-        lotteryLastRunAt: null,
-        lotteryStatus: null,
-      });
-
-      await expect(
-        async () =>
-          await service.lotteryGenerate(
-            { user: requestingUser } as unknown as ExpressRequest,
-            {} as unknown as Response,
-            { listingId },
-          ),
-      ).rejects.toThrowError(
-        `Listing ${listingId}: the lottery was attempted to be generated but it was already run previously`,
-      );
-
-      expect(prisma.listings.findUnique).toHaveBeenCalledWith({
-        select: {
-          id: true,
-          lotteryLastRunAt: true,
-          lotteryStatus: true,
-        },
-        where: {
-          id: listingId,
-        },
-      });
-
-      expect(prisma.applications.findMany).not.toHaveBeenCalled();
-
-      expect(
-        prisma.applicationLotteryPositions.createMany,
-      ).not.toHaveBeenCalled();
-      expect(prisma.listings.update).not.toHaveBeenCalled();
     });
   });
 });
